Chore: Split up into separate modules. Although, still a lot of work to be done as typing, docstrings and removal of properties on the classes.
There will probably be more modules created for other things such as
pickups, objects, etc.
